requirements.txt
Django==1.10.6
djangorestframework==3.5.3
httplib2==0.10.3
PySocks==1.6.7
pytz==2017.2
six==1.10.0
twilio==5.7.0
